refactor: extract project form logic
