+++ src/pages/Index.tsx
@@ -1,7 +1,5 @@
 import { useState, useEffect } from "react";
-import { useUser } from "@supabase/auth-helpers-react";
 import { useNavigate } from "react-router-dom";
-import { supabase } from "@/integrations/supabase/client";
 import { DashboardHeader } from "@/components/DashboardHeader";
 import { UserInfo } from "@/components/UserInfo";
 import { ViewMode } from "@/components/ViewToggle";
@@ -12,17 +10,12 @@ import { ProjectList } from "@/components/project/ProjectList";
 import { ProjectModals } from "@/components/project/ProjectModals";
 import { usePermissionsContext } from "@/contexts/PermissionsContext";
 import { useProjectsListView } from "@/hooks/use-projects-list-view";
-
-const LoadingSpinner = () => (
-  <div className="flex justify-center items-center h-48">
-    <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-gray-900"></div>
-  </div>
-);
+import { useProjectFormHandlers } from "@/hooks/useProjectFormHandlers";
+import { LoadingSpinner } from "@/components/ui/loading-spinner";
 
 const Index = () => {
   const navigate = useNavigate();
   const { isLoading: isPermissionsLoading, isError: isPermissionsError, userProfile } = usePermissionsContext();
-  const user = useUser();
 
   const [isProjectFormOpen, setIsProjectFormOpen] = useState(false);
   const [isProjectSelectionOpen, setIsProjectSelectionOpen] = useState(false);
@@ -32,6 +25,8 @@ const Index = () => {
     id: string;
     title: string;
   } | null>(null);
+
+  const { handleProjectFormSubmit } = useProjectFormHandlers(selectedProject);
   
   const [searchQuery, setSearchQuery] = useState(() => {
     return localStorage.getItem("projectSearchQuery") || "";
@@ -175,248 +170,6 @@ const Index = () => {
     setSelectedProject(null);
   };
 
-  const handleProjectFormSubmit = async (projectData: {
-    title: string;
-    description: string;
-    projectManager: string;
-    startDate?: Date;
-    endDate?: Date;
-    priority: string;
-    monitoringLevel: MonitoringLevel;
-    monitoringEntityId: string | null;
-    owner_id: string | null;
-    poleId: string;
-    directionId: string;
-    serviceId: string;
-    lifecycleStatus: ProjectLifecycleStatus;
-    for_entity_type: ForEntityType;
-    for_entity_id: string | undefined;
-    innovation: {
-      novateur: number;
-      usager: number;
-      ouverture: number;
-      agilite: number;
-      impact: number;
-    };
-    framing: {
-      context: string;
-      stakeholders: string;
-      governance: string;
-      objectives: string;
-      timeline: string;
-      deliverables: string;
-    };
-  }) => {
-    try {
-      // console.log("[ProjectForm] Starting project operation with detailed logs");
-      // console.log("[ProjectForm] Current user session:", await supabase.auth.getSession());
-      // console.log("[ProjectForm] User roles from context:", userProfile?.roles);
-
-      const projectPayload = {
-        title: projectData.title,
-        description: projectData.description,
-        project_manager: projectData.projectManager,
-        start_date: projectData.startDate?.toISOString().split('T')[0],
-        end_date: projectData.endDate?.toISOString().split('T')[0],
-        priority: projectData.priority,
-        owner_id: projectData.owner_id,
-        pole_id: projectData.poleId === "none" ? null : projectData.poleId,
-        direction_id: projectData.directionId === "none" ? null : projectData.directionId,
-        service_id: projectData.serviceId === "none" ? null : projectData.serviceId,
-        lifecycle_status: projectData.lifecycleStatus,
-        for_entity_type: projectData.for_entity_type,
-        for_entity_id: projectData.for_entity_id,
-      };
-
-      // console.log("[ProjectForm] Project payload:", projectPayload);
-
-      const { data: userRoles, error: rolesError } = await supabase
-        .from('user_roles')
-        .select('role')
-        .eq('user_id', user?.id);
-      
-      // console.log("[ProjectForm] User roles from direct query:", userRoles);
-      if (rolesError) {
-        console.error("[ProjectForm] Error fetching roles:", rolesError);
-      }
-
-      let projectId;
-
-      if (selectedProject?.id) {
-        // console.log("[ProjectForm] Updating project:", selectedProject.id);
-        const { data: updatedProject, error: projectError } = await supabase
-          .from("projects")
-          .update(projectPayload)
-          .eq("id", selectedProject.id)
-          .select();
-
-        // console.log("[ProjectForm] Project update result:", { updatedProject, error: projectError });
-
-        if (projectError) {
-          console.error("[ProjectForm] Project update error:", projectError);
-          throw projectError;
-        }
-
-        projectId = selectedProject.id;
-
-        // console.log("[ProjectForm] Updating innovation scores for project:", selectedProject.id);
-        const { data: innovationData, error: innovationError } = await supabase
-          .from("project_innovation_scores")
-          .upsert({
-            project_id: selectedProject.id,
-            ...projectData.innovation,
-            updated_at: new Date().toISOString()
-          }, {
-            onConflict: 'project_id'
-          })
-          .select();
-
-        // console.log("[ProjectForm] Innovation scores update result:", { innovationData, error: innovationError });
-
-        if (innovationError) {
-          console.error("[ProjectForm] Innovation scores update error:", innovationError);
-          throw innovationError;
-        }
-
-        // console.log("[ProjectForm] Updating monitoring for project:", selectedProject.id);
-        const { data: monitoringData, error: monitoringError } = await supabase
-          .from("project_monitoring")
-          .upsert({
-            project_id: selectedProject.id,
-            monitoring_level: projectData.monitoringLevel,
-            monitoring_entity_id: projectData.monitoringEntityId,
-          }, {
-            onConflict: 'project_id'
-          })
-          .select();
-
-        // console.log("[ProjectForm] Monitoring update result:", { monitoringData, error: monitoringError });
-
-        if (monitoringError) {
-          console.error("[ProjectForm] Monitoring update error:", monitoringError);
-          throw monitoringError;
-        }
-
-      } else {
-        // console.log("[ProjectForm] Creating new project with user:", user?.id);
-        const { data: newProject, error: projectError } = await supabase
-          .from("projects")
-          .insert(projectPayload)
-          .select()
-          .single();
-
-        // console.log("[ProjectForm] Project creation result:", { newProject, error: projectError });
-
-        if (projectError) {
-          console.error("[ProjectForm] Project creation error:", projectError.message);
-          console.error("[ProjectForm] Project creation error details:", projectError);
-          throw projectError;
-        }
-
-        projectId = newProject.id;
-
-        // console.log("[ProjectForm] Creating innovation scores for new project:", newProject.id);
-        const { data: innovationData, error: innovationError } = await supabase
-          .from("project_innovation_scores")
-          .insert({
-            project_id: newProject.id,
-            ...projectData.innovation
-          })
-          .select();
-
-        // console.log("[ProjectForm] Innovation scores creation result:", { innovationData, error: innovationError });
-
-        if (innovationError) {
-          console.error("[ProjectForm] Innovation scores creation error:", innovationError.message);
-          console.error("[ProjectForm] Innovation scores creation error details:", innovationError);
-          throw innovationError;
-        }
-
-        // console.log("[ProjectForm] Creating monitoring for new project:", newProject.id);
-        const { data: monitoringData, error: monitoringError } = await supabase
-          .from("project_monitoring")
-          .insert({
-            project_id: newProject.id,
-            monitoring_level: projectData.monitoringLevel,
-            monitoring_entity_id: projectData.monitoringEntityId,
-          })
-          .select();
-
-        // console.log("[ProjectForm] Monitoring creation result:", { monitoringData, error: monitoringError });
-
-        if (monitoringError) {
-          console.error("[ProjectForm] Monitoring creation error:", monitoringError.message);
-          console.error("[ProjectForm] Monitoring creation error details:", monitoringError);
-          throw monitoringError;
-        }
-      }
-
-      if (projectId) {
-        // console.log("[ProjectForm] Checking existing framing data for project:", projectId);
-        
-        const { data: existingFraming, error: checkFramingError } = await supabase
-          .from("project_framing")
-          .select("id")
-          .eq("project_id", projectId)
-          .maybeSingle();
-          
-        // console.log("[ProjectForm] Existing framing check result:", { existingFraming, error: checkFramingError });
-        
-        if (checkFramingError) {
-          console.error("[ProjectForm] Error checking existing framing:", checkFramingError);
-          throw checkFramingError;
-        }
-
-        const framingData = {
-          project_id: projectId,
-          context: projectData.framing.context,
-          stakeholders: projectData.framing.stakeholders,
-          governance: projectData.framing.governance,
-          objectives: projectData.framing.objectives,
-          timeline: projectData.framing.timeline,
-          deliverables: projectData.framing.deliverables,
-        };
-
-        if (existingFraming) {
-          // console.log("[ProjectForm] Updating existing framing data for project:", projectId);
-          const { data: updatedFraming, error: updateFramingError } = await supabase
-            .from("project_framing")
-            .update(framingData)
-            .eq("id", existingFraming.id)
-            .select();
-
-          // console.log("[ProjectForm] Framing update result:", { updatedFraming, error: updateFramingError });
-          
-          if (updateFramingError) {
-            console.error("[ProjectForm] Error updating framing data:", updateFramingError);
-            throw updateFramingError;
-          }
-        } else {
-          // console.log("[ProjectForm] Creating new framing data for project:", projectId);
-          const { data: newFraming, error: insertFramingError } = await supabase
-            .from("project_framing")
-            .insert(framingData)
-            .select();
-
-          // console.log("[ProjectForm] Framing creation result:", { newFraming, error: insertFramingError });
-          
-          if (insertFramingError) {
-            console.error("[ProjectForm] Error inserting framing data:", insertFramingError);
-            throw insertFramingError;
-          }
-        }
-      } else {
-        console.error("[ProjectForm] No project ID available to save framing data");
-      }
-
-      // console.log("[ProjectForm] Project operation completed successfully");
-      
-      return { id: projectId };
-    } catch (error) {
-      console.error("[ProjectForm] Operation failed:", error);
-      throw error;
-    }
-  };
 
   const handleNewReview = () => {
     setIsProjectSelectionOpen(true);
